Move http recorder to an internal library
No functional change, just moving code
