+++ internal/logger/audit.go
@@ -18,10 +18,8 @@
 package logger
 
 import (
-	"bytes"
 	"context"
 	"fmt"
-	"io"
 	"net/http"
 	"strconv"
 	"time"
@@ -29,110 +27,9 @@ import (
 	"github.com/klauspost/compress/gzhttp"
 	"github.com/minio/madmin-go"
 	xhttp "github.com/minio/minio/internal/http"
+	httplogger "github.com/minio/minio/internal/http/logger"
 	"github.com/minio/minio/internal/logger/message/audit"
 )
-
-// ResponseWriter - is a wrapper to trap the http response status code.
-type ResponseWriter struct {
-	http.ResponseWriter
-	StatusCode int
-	// Log body of 4xx or 5xx responses
-	LogErrBody bool
-	// Log body of all responses
-	LogAllBody bool
-
-	TimeToFirstByte time.Duration
-	StartTime       time.Time
-	// number of bytes written
-	bytesWritten int
-	// number of bytes of response headers written
-	headerBytesWritten int
-	// Internal recording buffer
-	headers bytes.Buffer
-	body    bytes.Buffer
-	// Indicate if headers are written in the log
-	headersLogged bool
-}
-
-// NewResponseWriter - returns a wrapped response writer to trap
-// http status codes for auditing purposes.
-func NewResponseWriter(w http.ResponseWriter) *ResponseWriter {
-	return &ResponseWriter{
-		ResponseWriter: w,
-		StatusCode:     http.StatusOK,
-		StartTime:      time.Now().UTC(),
-	}
-}
-
-func (lrw *ResponseWriter) Write(p []byte) (int, error) {
-	if !lrw.headersLogged {
-		// We assume the response code to be '200 OK' when WriteHeader() is not called,
-		// that way following Golang HTTP response behavior.
-		lrw.WriteHeader(http.StatusOK)
-	}
-	n, err := lrw.ResponseWriter.Write(p)
-	lrw.bytesWritten += n
-	if lrw.TimeToFirstByte == 0 {
-		lrw.TimeToFirstByte = time.Now().UTC().Sub(lrw.StartTime)
-	}
-	if (lrw.LogErrBody && lrw.StatusCode >= http.StatusBadRequest) || lrw.LogAllBody {
-		// Always logging error responses.
-		lrw.body.Write(p)
-	}
-	if err != nil {
-		return n, err
-	}
-	return n, err
-}
-
-// Write the headers into the given buffer
-func (lrw *ResponseWriter) writeHeaders(w io.Writer, statusCode int, headers http.Header) {
-	n, _ := fmt.Fprintf(w, "%d %s\n", statusCode, http.StatusText(statusCode))
-	lrw.headerBytesWritten += n
-	for k, v := range headers {
-		n, _ := fmt.Fprintf(w, "%s: %s\n", k, v[0])
-		lrw.headerBytesWritten += n
-	}
-}
-
-// BodyPlaceHolder returns a dummy body placeholder
-var BodyPlaceHolder = []byte("<BODY>")
-
-// Body - Return response body.
-func (lrw *ResponseWriter) Body() []byte {
-	// If there was an error response or body logging is enabled
-	// then we return the body contents
-	if (lrw.LogErrBody && lrw.StatusCode >= http.StatusBadRequest) || lrw.LogAllBody {
-		return lrw.body.Bytes()
-	}
-	// ... otherwise we return the <BODY> place holder
-	return BodyPlaceHolder
-}
-
-// WriteHeader - writes http status code
-func (lrw *ResponseWriter) WriteHeader(code int) {
-	if !lrw.headersLogged {
-		lrw.StatusCode = code
-		lrw.writeHeaders(&lrw.headers, code, lrw.ResponseWriter.Header())
-		lrw.headersLogged = true
-		lrw.ResponseWriter.WriteHeader(code)
-	}
-}
-
-// Flush - Calls the underlying Flush.
-func (lrw *ResponseWriter) Flush() {
-	lrw.ResponseWriter.(http.Flusher).Flush()
-}
-
-// Size - returns  the number of bytes written
-func (lrw *ResponseWriter) Size() int {
-	return lrw.bytesWritten
-}
-
-// HeaderSize - returns the number of bytes of response headers written
-func (lrw *ResponseWriter) HeaderSize() int {
-	return lrw.headerBytesWritten
-}
 
 const contextAuditKey = contextKeyType("audit-entry")
 
@@ -197,13 +94,13 @@ func AuditLog(ctx context.Context, w http.ResponseWriter, r *http.Request, reqCl
 			headerBytes     int64
 		)
 
-		var st *ResponseWriter
+		var st *httplogger.ResponseRecorder
 		switch v := w.(type) {
-		case *ResponseWriter:
+		case *httplogger.ResponseRecorder:
 			st = v
 		case *gzhttp.GzipResponseWriter:
 			// the writer may be obscured by gzip response writer
-			if rw, ok := v.ResponseWriter.(*ResponseWriter); ok {
+			if rw, ok := v.ResponseWriter.(*httplogger.ResponseRecorder); ok {
 				st = rw
 			}
 		}
